made name change process
