--- frontend/src/commands/src/types/mapper.d.ts
@@ -1,22 +1,0 @@
-// 変数名の変換文字列の定義
-export interface FirstQuestion {
-interviewId: string;
-question: string;
-}
-
-export interface NextQuestion {
-questionId: number;
-question: string;
-}
-
-export interface Feedback {
-questionId: number;
-response: string;
-score: number;
-continueQuestion: boolean;
-}
-
-export interface GeneralFeedback {
-scores: number[];
-generalReview: string;
-}
