+++ frontend/src/webview/src/screens/answer-screen/context/AnswerContextProvider.tsx
@@ -2,12 +2,9 @@ import { useState, useCallback, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import type { ChatMessage } from "@/types/chat-message";
 import type { Character } from "@/types/character";
-import {
-  FeedBackResponse,
-  GeneralFeedbackResponse,
-} from "@shared/backend-api-response-value";
 import type { VscodeApiRequestValue } from "@shared/vscode-api-request-value";
 import type { VscodeApiResponseValue } from "@shared/vscode-api-response-value";
+import { Feedback, GeneralFeedback } from "@shared/webview-api-response-type";
 import { useLoading } from "@/screens/context/LoadingContext";
 import { useThinkingAnimation } from "@/screens/components/hooks/use-thinking-animation";
 import { AnswerContext } from "@/screens/answer-screen/context/UseAnswerContext";
@@ -43,10 +40,10 @@ export const AnswerContextProvider: React.FC<AnswerContextProviderProps> = ({
   const navigate = useNavigate();
 
   const judgeContinueSameQuestion = useCallback(
-    (payload: FeedBackResponse) => {
+    (payload: Feedback) => {
       stopThinking();
       const lastScore = payload.score;
-      if (payload.continue_question) {
+      if (payload.continueQuestion) {
         setChatHistory((prev) => [
           ...prev,
           { type: "question", text: payload.response },
@@ -54,7 +51,7 @@ export const AnswerContextProvider: React.FC<AnswerContextProviderProps> = ({
         setDisplayEnterBox(true);
       } else {
         const total = currentCharacter.totalQuestion;
-        const lastId = payload.question_id;
+        const lastId = payload.questionId;
         setButtonDisplay(lastId >= total ? "最終結果へ" : "次へ");
         setChatHistory((prev) => [
           ...prev,
@@ -73,8 +70,8 @@ export const AnswerContextProvider: React.FC<AnswerContextProviderProps> = ({
     const msg: VscodeApiRequestValue = {
       type: "fetchFeedback",
       payload: {
-        interview_id: interviewId,
-        question_id: questionId,
+        interviewId: interviewId,
+        questionId: questionId,
         answer: chatInput,
       },
     };
@@ -87,7 +84,7 @@ export const AnswerContextProvider: React.FC<AnswerContextProviderProps> = ({
     const nextQuestionId = questionId + 1;
     const msg: VscodeApiRequestValue = {
       type: "fetchNextQuestion",
-      payload: { interview_id: interviewId, question_id: nextQuestionId },
+      payload: { interviewId: interviewId, questionId: nextQuestionId },
     };
     vscode.postMessage(msg);
     setQuestionId(nextQuestionId);
@@ -99,13 +96,13 @@ export const AnswerContextProvider: React.FC<AnswerContextProviderProps> = ({
     showLoading("全ての回答をチェック中・・・");
     const msg: VscodeApiRequestValue = {
       type: "fetchGeneralFeedback",
-      payload: { interview_id: interviewId },
+      payload: { interviewId: interviewId },
     };
     vscode.postMessage(msg);
   }, [showLoading, vscode, interviewId]);
 
   const moveGeneralFeedbackScreen = useCallback(
-    (payload: GeneralFeedbackResponse) => {
+    (payload: GeneralFeedback) => {
       navigate("/feedback", {
         state: {
           payload,
@@ -129,7 +126,7 @@ export const AnswerContextProvider: React.FC<AnswerContextProviderProps> = ({
             ...prev,
             { type: "question", text: msg.payload.question },
           ]);
-          setQuestionId(msg.payload.question_id);
+          setQuestionId(msg.payload.questionId);
           setButtonDisplay("スキップ");
           setDisplayEnterBox(true);
           break;
